requirements.txt
alembic==1.15.1
Faker==37.0.0
Mako==1.3.9
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
python-dotenv==1.0.1
SQLAlchemy==2.0.39
typing_extensions==4.12.2
tzdata==2025.1
